requirements.txt
gradio==5.49.0
requests==2.32.5
openai==2.1.0
python-dotenv==1.1.1
pandas==2.3.3
beautifulsoup4==4.14.2
chromadb==1.1.0